Word to guess is now shown
Added underscores that match the number of letters in the word to be guessed. These underscores change to the appropriate guessed letter when a correct guess is made. Changed the guess update to allow for showing all instances of the letter guessed in the word when correctly guessed, rather than only the first instance.